requirements.txt
cvxpy==1.6.4
ecos==2.0.14
filterpy==1.4.5
google-generativeai==0.8.4
imageio==2.36.1
imageio-ffmpeg==0.6.0
ipython==8.12.3
joblib==1.3.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
Mosek==10.2.15
moviepy==1.0.3
numpy==1.26.4
opencv-python==4.11.0.86
pandas==2.1.4
PyYAML==6.0.2
scenedetect==0.6.6
streamlit==1.41.1
torch==2.2.1
tqdm==4.67.1
ultralytics==8.3.91
yt-dlp==2025.2.19
